Modify: MyPage의 이미지 요청 함수들을 MyPictures로 이동
